refactor: centralizes route module handling
Moves route module types and imports to dedicated module file

Eliminates duplicate import.meta.glob calls across components by consolidating route module handling into a single location

Improves code organization and reduces redundancy
