+++ src/lib/components/Head.svelte
@@ -1,11 +1,7 @@
 <script lang="ts">
 	import { page } from "$app/state";
 	import { defaultDescription, defaultImage, defaultTitle } from "$lib/constants";
-	import type { ModuleData } from "$lib/util";
-
-	const routeModules: Record<string, ModuleData> = import.meta.glob("/src/routes/**/+page.svelte", {
-		eager: true
-	});
+	import { type ModuleData, routeModules } from "$lib/modules";
 
 	function getHeadFromModule(module: ModuleData | undefined) {
 		if (module?.pageHead) return module.pageHead;
